// apps/backend-api/src/modules/audit/enums/audit-action.enum.ts
export enum AuditAction {
  LOGIN_SUCCESS = 'LOGIN_SUCCESS',
  LOGIN_FAILED = 'LOGIN_FAILED',
  LOGOUT = 'LOGOUT',
  TOKEN_REFRESH = 'TOKEN_REFRESH',
  OTP_VERIFIED = 'OTP_VERIFIED',
  PASSWORD_CHANGED = 'PASSWORD_CHANGED',
  ACCOUNT_LOCKED = 'ACCOUNT_LOCKED',

  PRODUCT_CREATED = 'PRODUCT_CREATED',
  PRODUCT_UPDATED = 'PRODUCT_UPDATED',
  PRODUCT_DELETED = 'PRODUCT_DELETED',
  PRODUCT_IMAGE_UPLOADED = 'PRODUCT_IMAGE_UPLOADED',

  CATEGORY_CREATED = 'CATEGORY_CREATED',
  CATEGORY_UPDATED = 'CATEGORY_UPDATED',
  CATEGORY_DELETED = 'CATEGORY_DELETED',

  USER_BLOCKED = 'USER_BLOCKED',
  USER_UNBLOCKED = 'USER_UNBLOCKED',

  PERMISSION_GRANTED = 'PERMISSION_GRANTED',
  PERMISSION_REVOKED = 'PERMISSION_REVOKED',
  PERMISSIONS_SEEDED = 'PERMISSIONS_SEEDED',

  ORDER_STATUS_CHANGED = 'ORDER_STATUS_CHANGED',
  COUPON_APPLIED = 'COUPON_APPLIED',

  PAYMENT_CREATED = 'PAYMENT_CREATED',
}